requirements.txt
aiofiles==23.2.1
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
distlib==0.3.7
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.101.1
fastapi-mail==1.4.1
filelock==3.12.3
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.5
passlib==1.7.4
Pillow==10.0.1
pipenv==2023.9.1
platformdirs==3.10.0
pydantic==2.2.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.3
pydantic_core==2.6.1
PyJWT==2.8.0
pymongo==3.11.0
pypika-tortoise==0.1.6
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.4
watchfiles==0.19.0
websockets==11.0.3
